frontend: Separate plugins/index.tsx into index.ts and lib.ts
The lib.ts file should carry the bits to be used by plugins whereas
the index.ts should have the functions that Headlamp itself needs for
loading the plugins.
