+++ plugins/headlamp-plugin/lib/index.ts
@@ -1,7 +1,7 @@
 import * as K8s from '../types/lib/k8s';
 import * as ApiProxy from '../types/lib/k8s/apiProxy';
 import * as Utils from '../types/lib/util';
-import { Headlamp, Plugin } from '../types/plugin/index';
+import { Headlamp, Plugin } from '../types/plugin/lib';
 import Registry from '../types/plugin/registry';
 import * as CommonComponents from './CommonComponents';
